:wrench: fix: wishlist delete method

+++ src/wishlist/wishlist.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, Post } from "@nestjs/common";
+import { Body, Controller, Delete, Get, Param, Post, Query } from "@nestjs/common";
 import { WishListService } from "./wishlist.service";
 import { CreateWishListDTO } from "./dto/create-wishlist.dto";
 import { Public } from "src/auth/decorator/public.decorator";
@@ -14,7 +14,9 @@ export class WishListController {
     }
 
     @Delete()
-    async remove(@Body() userId: number, productId: number) {
-        return await this.wishListService.remove({ userId, productId } as CreateWishListDTO);
+    async remove(@Query("userId") userId: number, @Query("productId") productId: number) {
+        const parsedUserId = Number(userId);
+        const parsedProductId = Number(productId);
+        return await this.wishListService.remove({ userId: parsedUserId, productId: parsedProductId } as CreateWishListDTO);
     }
 }
